Allow to use a local SSD for executors firecracker data dir (#82)
## Why local SSDs?

This PR adds support for mounting an SSD from the host directly, as opposed to having a larger boot SSD disk over iSCSI. A local SSD has higher bandwidth, lower latency, and much higher IOPS. This benefits the executors a lot - IO has always been the big bottleneck if doing multi-concurrency on a single host. Grafana says this new disk is about half as busy as the previous attempt.

*But we only need ~20GB plus collateral for each job 😢 😢*
True! But - and here comes the great part about this disk type:
Local SSDs are the only disk type (afaik) that support spot pricing as well - and we use spot instances :) 
Discounted price per GB for `ssd-pd`: $0.1581
Discounted price per GB for local SSD attached to spot instance: $0.04464

So one local SSD costs about $16.74 per month.
For the same price, we get 105.8GB of `ssd-pd` storage.
Guess what? We used to give the machines 100GB! Not because we need exactly 100GB, but because we needed the additional IOPS (which increase with every GB you allocate).
So this should be +- 0, at well improved performance. I've also seen this much more stable at higher concurrency per host - so potentially we could save even more.

Side-note: For the Cloud cluster, we always used much bigger disks, because we wanted auto indexing there to be as fast as possible. So on Cloud alone, we should save a good bunch. Also, this makes sure we use the same perf characteristics on both our largest instance and on managed instances, since we don't override this anymore.

### Performance comparison:

For a C2 instance with 8 VCPUs (our default) we get:

max 4000 IOPS read, 4000 IOPS write, 240MBps read, 240 MBps write
But per GB we allocate, we only get 30 IOPS, and 0.48 MBps of throughput.
At the 100 GB we use by default, that means merely 48MBps of throughput and 3000 IOPS.

With a local SSD, however, we get:
170,000 IOPS read, 90,000 IOPS write, and 660 MBps read, 350 MBps write.

### Question

We've arbitrarily picked 20GB as the maximum size for the firecracker VM disk - we might as well increase that default to 50GB and make executors work for even larger mono repos. This would incur zero additional cost for us when using local SSDs, if we go with local SSDs for MIs, too. 

### Test plan

Ran over the weekend on dotcom, no issues there.