Potentially fix issue with boot ordering (#179)
It seems that recently docker started to run after the iptables-restore
command, which then fails because it depends on some things to be
created by docker I think.

This might fix our immediate issue by loading the chain manually again
after boot.

Test plan: We'll need to upgrade an instance to use this. But running it
manually on a broken one does make it start the executor service
afterwards.